Consolidate some things
