+++ src/worker/browser/worker_connection.ts
@@ -23,7 +23,7 @@
 
 /* eslint-disable no-console */
 import * as vscode from "vscode";
-import { fetchFile } from "../../extension/browser/utils";
+import { fetchFile } from "../../extension/utils";
 import { Message, WorkerMessage, WorkerReadMessage } from "../types";
 const workerLog = vscode.window.createOutputChannel("Malloy Worker");
 // const debugWorker = process.env.MALLOY_DEBUG_WORKER === "true";
